refactor(solver): split call inference guard helpers
